Reordered audio effects chain, added new effects
The audio effects chain was reordered to be more logical. Effects are
applied in the order they are written: it makes more sense to apply
reverb _after_ applying the low pass filter to avoid the reverberations
being filtered. Optional audio effects are now applied in an order that
will not break the effects chain. The command line arguments were put
into new groups to make the help text more legible.